+++ src/domain/entities/Conversation.ts
@@ -8,13 +8,3 @@ export interface Conversation {
   userId: number;
   topicId: number | null;
 }
-
-export interface User {
-  id: number;
-  username: string;
-}
-
-export interface Topic {
-  id: number;
-  name: string;
-}
