+++ src/domain/entities/Topic.ts
@@ -1,0 +1,4 @@
+export interface Topic {
+  id: number;
+  name: string;
+}
