+++ src/domain/entities/User.ts
@@ -1,0 +1,4 @@
+export interface User {
+  id: number;
+  username: string;
+}
